safari-534.53-branch/LayoutTests/platform/qt/fast/text/find-soft-hyphen-expected.txt

Tests find for strings with soft hyphens in them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL canFind("ab", "a" + softHyphen + "b") should be true. Was false.
FAIL canFind("ab", "a" + softHyphen + softHyphen + "b") should be true. Was false.
FAIL canFind("àb", "a" + softHyphen + "b") should be true. Was false.
FAIL canFind("ab", "a" + softHyphen + "̀b") should be true. Was false.
FAIL canFind("ab", "à" + softHyphen + "b") should be true. Was false.
PASS canFind("a" + softHyphen + "b", "a" + softHyphen + "b") is true
FAIL canFind("a" + softHyphen + "b", "ab") should be true. Was false.
PASS canFind("a" + hyphen + "b", "a" + softHyphen + "b") is false
PASS canFind("a-b", "a" + softHyphen + "b") is false
PASS successfullyParsed is true

TEST COMPLETE
